+++ src/webview/ui/listeners/index.ts
@@ -4,6 +4,7 @@ import { UpdateRatio, UpdateTargetProportion } from "../updaters";
 import { addButtonListeners } from "./addButtonListeners";
 import { addContentEditableListeners } from "./addContentEditableListeners";
 import { addHideOneTimeConfigListener } from "./addHideOneTimeConfigListener";
+import { addNumberInputListeners } from "./addNumberInputListeners";
 import { addOpenConfigListeners } from "./addOpenConfigListeners";
 import { addSelectListeners } from "./addSelectListeners";
 import { addToogleListeners } from "./addToggleListeners";
@@ -17,6 +18,7 @@ export function addListeners() {
     addContentEditableListeners();
     addHideOneTimeConfigListener();
     addOpenConfigListeners();
+    addNumberInputListeners();
 
     const resizeObserver = new ResizeObserver(() => {
         if (SessionConfig.get("aspectRatio") !== "none") {